refactor: remove attack button replacement and rely on core card creation logic
