+++ module.json
@@ -2,7 +2,7 @@
   "name": "mre-dnd5e",
   "title": "Minimal Rolling Enhancements for D&D5e",
   "description": "Some minimalist enhancements to the core D&D5e rolling workflow. Attempts to stay as close to core as possible while improving convenience.",
-  "version": "2.0.1",
+  "version": "3.0.0",
   "author": "Cole Schultz, Andrew Krigline",
   "authors": [
     {
